+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'TenThread'
+rootProject.name = 'tenThread'
