+++ src/main/java/com/example/tenthread/dto/ProfilePasswordCheckDto.java
@@ -1,0 +1,10 @@
+package com.example.tenthread.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ProfilePasswordCheckDto {
+    private String password;
+}
