+++ src/main/java/com/example/tenthread/service/UserService.java
@@ -1,9 +1,6 @@
 package com.example.tenthread.service;
 
-import com.example.tenthread.dto.LoginRequestDto;
-import com.example.tenthread.dto.ProfileRequestDto;
-import com.example.tenthread.dto.UserRequestDto;
-import com.example.tenthread.dto.UserResponseDto;
+import com.example.tenthread.dto.*;
 import com.example.tenthread.entity.User;
 import com.example.tenthread.entity.UserRoleEnum;
 import com.example.tenthread.jwt.JwtUtil;
@@ -58,12 +55,19 @@ public class UserService {
         response.addHeader("Authorization", token);
     }
 
-    public void beforeProfilePasswordCheck(User user, String password) {
+    /**
+     * 프로필 GET 페이지로 가기 전에 비밀번호 확인
+     * @param user  : 로그인 한 유저
+     * @param profilePasswordCheckDto : 입력한 패스워드
+     */
+    public void beforeProfilePasswordCheck(User user, ProfilePasswordCheckDto profilePasswordCheckDto) {
         User beforeProfileUser = userRepository.findByUsername(user.getUsername()).orElseThrow(
                 () -> new IllegalArgumentException("로그인해주세요.")
         );
 
-        if(!passwordEncoder.matches(password, beforeProfileUser.getPassword())) {
+        String checkPass = profilePasswordCheckDto.getPassword();
+
+        if(!passwordEncoder.matches(checkPass, beforeProfileUser.getPassword())) {
             throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
         }
     }
@@ -86,7 +90,7 @@ public class UserService {
         String newPassword = profileRequestDto.getPassword();
 
         updateProfile.setNickname(newNickname);
-        updateProfile.setPassword(newPassword);
+        updateProfile.setPassword(passwordEncoder.encode(newPassword));
 
         userRepository.save(updateProfile);
     }
